Revert "shamu: Disable forced encryption"
This is causing some weird issues with the /data partition
when it is encrypted coming from stock.  TWRP's behavior keeps encryption
enabled when it does a data wipe. Users would have to manually format data
in a separate process when coming from encrypted stock.

The CDD also explains that force encryption for L devices is "very
strongly RECOMMENDED and we expect this to change to MUST in future
versions of Android."

For these reasons we will leave forced encryption enabled

This reverts commit 49aa01d0827e92caf9eee87a89220ee2f5a43ade.

Change-Id: I4c66c11b256b75ec35570d4aa959b4a17a560cf9